Theme change on IE works
