+++ packages/base/src/Theming.js
@@ -1,8 +1,7 @@
 import { addCustomCSS, getCustomCSS } from "./theming/CustomStyle.js";
 import { getThemeProperties } from "./theming/ThemeProperties.js";
-import { injectThemeProperties } from "./theming/StyleInjection.js";
+import injectThemeProperties from "./theming/injectThemeProperties.js";
 
-const defaultTheme = "sap_fiori_3";
 const themeChangeCallbacks = [];
 
 const attachThemeChange = function attachThemeChange(callback) {
@@ -12,17 +11,13 @@ const attachThemeChange = function attachThemeChange(callback) {
 };
 
 const _applyTheme = async theme => {
-	let cssText = "";
-
-	if (theme !== defaultTheme) {
-		cssText = await getThemeProperties("@ui5/webcomponents", theme);
-	}
-	injectThemeProperties(cssText);
-	_executeThemeChangeCallbacks(theme);
+	const cssText = await getThemeProperties("@ui5/webcomponents", theme);
+	const vars = injectThemeProperties(cssText);
+	_executeThemeChangeCallbacks(theme, vars);
 };
 
-const _executeThemeChangeCallbacks = theme => {
-	themeChangeCallbacks.forEach(callback => callback(theme));
+const _executeThemeChangeCallbacks = (theme, vars) => {
+	themeChangeCallbacks.forEach(callback => callback(theme, vars));
 };
 
 const getEffectiveStyle = ElementClass => {
