+++ packages/main/src/ThemePropertiesProvider.js
@@ -2,6 +2,8 @@ import { registerThemeProperties } from "@ui5/webcomponents-base/dist/theming/Th
 
 import belizeThemeProperties from "./generated/themes/sap_belize/parameters-bundle.css.js";
 import belizeHcbThemeProperties from "./generated/themes/sap_belize_hcb/parameters-bundle.css.js";
+import fiori3ThemeProperties from "./generated/themes/sap_fiori_3/parameters-bundle.css.js";
 
+registerThemeProperties("@ui5/webcomponents", "sap_fiori_3", fiori3ThemeProperties);
 registerThemeProperties("@ui5/webcomponents", "sap_belize", belizeThemeProperties);
 registerThemeProperties("@ui5/webcomponents", "sap_belize_hcb", belizeHcbThemeProperties);
